Initial version with some spring boot data
