+++ pom.xml
@@ -55,6 +55,11 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+      		<groupId>org.locationtech.jts</groupId>
+      		<artifactId>jts-core</artifactId>
+      		<version>1.19.0</version>
+    	</dependency>
 	</dependencies>
 
 	<build>
