+++ src/main/java/com/datastax/oss/cass_stac/App.java
@@ -1,1075 +1,13 @@
 package com.datastax.oss.cass_stac;
 
-import com.datastax.oss.cass_stac.dao.DaoFactory;
-import com.datastax.oss.cass_stac.dao.FeatureCollectionDao;
-import com.datastax.oss.cass_stac.dao.ItemDao;
-import com.datastax.oss.cass_stac.model.FeatureCollection;
-import com.datastax.oss.cass_stac.model.Item;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+@SpringBootApplication
 public class App {
-    private static final Logger logger = LoggerFactory.getLogger(App.class);
-    static ObjectMapper objectMapper = new ObjectMapper();
 
-    public static void main(String[] args) throws Exception {
+	public static void main(String[] args) {
+		SpringApplication.run(App.class, args);
+	}
 
-        DaoFactory df = DaoFactory.getInstance();
-        logger.info("Hello World!");
-
-        logger.info("Saving Item...");
-        ItemDao itemDao = df.getDao(DaoFactory.DaoType.ITEM);
-
-        Item item = objectMapper.readValue(sampleItemJson(), Item.class);
-        itemDao.save(item);
-        logger.info("Done saving Item.");
-
-        logger.info("Retrieving Item...");
-        Item dbItem = itemDao.get("867ee240fffffff-2020-M12", "20201211_223832_CS2");
-        if (item.equals(dbItem))
-            logger.info("Items match!");
-        else
-            logger.error("Items do not match!");
-        logger.info("Done retrieving Item...");
-
-        logger.info("Saving Collection...");
-        FeatureCollectionDao featureCollectionDao = df.getDao(DaoFactory.DaoType.FEATURE_COLLECTION);
-        String sampleFeatureCollectionString = sampleFeatureCollectionJson();
-        FeatureCollection featureCollection = objectMapper.readValue(sampleFeatureCollectionString, FeatureCollection.class);
-        featureCollectionDao.save(featureCollection);
-        logger.info("Done saving Collection.");
-
-        System.exit(0);
-    }
-
-    private static String sampleItemJson() {
-        String jsonBase = """
-                {
-                  "stac_version": "1.0.0",
-                  "stac_extensions": [],
-                  "type": "Feature",
-                  "id": "20201211_223832_CS2",
-                  "bbox": [
-                    172.91173669923782,
-                    1.3438851951615003,
-                    172.95469614953714,
-                    1.3690476620161975
-                  ],
-                  "geometry": {
-                    "type": "Polygon",
-                    "coordinates": [
-                      [
-                        [
-                          172.91173669923782,
-                          1.3438851951615003
-                        ],
-                        [
-                          172.95469614953714,
-                          1.3438851951615003
-                        ],
-                        [
-                          172.95469614953714,
-                          1.3690476620161975
-                        ],
-                        [
-                          172.91173669923782,
-                          1.3690476620161975
-                        ],
-                        [
-                          172.91173669923782,
-                          1.3438851951615003
-                        ]
-                      ]
-                    ]
-                  },
-                  "properties": {
-                    "title": "Core Item",
-                    "description": "A sample STAC Item that includes examples of all common metadata",
-                    "start_datetime": "2020-12-11T22:38:32.125Z",
-                    "end_datetime": "2020-12-11T22:38:32.327Z",
-                    "created": "2020-12-12T01:48:13.725Z",
-                    "updated": "2020-12-12T01:48:13.725Z",
-                    "platform": "cool_sat1",
-                    "eo:cloud_cover": 13.1,
-                    "instruments": [
-                      "cool_sensor_v1"
-                    ],
-                    "constellation": "ion",
-                    "mission": "collection 5624",
-                    "gsd": 0.512
-                  },
-                  "collection": "simple-collection",""";
-        String jsonLinks = """
-                [
-                  {
-                    "rel": "collection",
-                    "href": "./collection.json",
-                    "type": "application/json",
-                    "title": "Simple Example Collection"
-                  },
-                  {
-                    "rel": "root",
-                    "href": "./collection.json",
-                    "type": "application/json",
-                    "title": "Simple Example Collection"
-                  },
-                  {
-                    "rel": "parent",
-                    "href": "./collection.json",
-                    "type": "application/json",
-                    "title": "Simple Example Collection"
-                  },
-                  {
-                    "rel": "alternate",
-                    "type": "text/html",
-                    "href": "http://remotedata.io/catalog/20201211_223832_CS2/index.html",
-                    "title": "HTML version of this STAC Item"
-                  }
-                ],""";
-        String jsonAssets = """
-                  {
-                    "analytic": {
-                      "href": "https://storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2_analytic.tif",
-                      "type": "image/tiff; application=geotiff; profile=cloud-optimized",
-                      "title": "4-Band Analytic",
-                      "roles": [
-                        "data"
-                      ]
-                    },
-                    "thumbnail": {
-                      "href": "https://storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg",
-                      "title": "Thumbnail",
-                      "type": "image/png",
-                      "roles": [
-                        "thumbnail"
-                      ]
-                    },
-                    "visual": {
-                      "href": "https://storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.tif",
-                      "type": "image/tiff; application=geotiff; profile=cloud-optimized",
-                      "title": "3-Band Visual",
-                      "roles": [
-                        "visual"
-                      ]
-                    },
-                    "udm": {
-                      "href": "https://storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2_analytic_udm.tif",
-                      "title": "Unusable Data Mask",
-                      "type": "image/tiff; application=geotiff;"
-                    },
-                    "json-metadata": {
-                      "href": "http://remotedata.io/catalog/20201211_223832_CS2/extended-metadata.json",
-                      "title": "Extended Metadata",
-                      "type": "application/json",
-                      "roles": [
-                        "metadata"
-                      ]
-                    },
-                    "ephemeris": {
-                      "href": "http://cool-sat.com/catalog/20201211_223832_CS2/20201211_223832_CS2.EPH",
-                      "title": "Satellite Ephemeris Metadata"
-                    }
-                  }
-                }
-                """;
-
-        return jsonBase + "\"links\": " + jsonLinks + "\"assets\": " + jsonAssets;
-    }
-
-    private static String sampleFeatureCollectionJson() {
-        String feature1 = """
-                                {
-                    "type": "Feature",
-                    "geometry": {
-                        "type": "Polygon",
-                        "coordinates": [
-                            [
-                                [
-                                    51.545876765758074,
-                                    25.10163158541922
-                                ],
-                                [
-                                    51.545876721263205,
-                                    25.101621561500913
-                                ],
-                                [
-                                    51.545887685270266,
-                                    25.1016114970669
-                                ],
-                                [
-                                    51.54587663227354,
-                                    25.101601513664257
-                                ],
-                                [
-                                    51.54587640979978,
-                                    25.101551394072377
-                                ],
-                                [
-                                    51.54585430382092,
-                                    25.101531427264177
-                                ],
-                                [
-                                    51.545843295326115,
-                                    25.101531467777274
-                                ],
-                                [
-                                    51.54583224234025,
-                                    25.10152148437103
-                                ],
-                                [
-                                    51.54584320634226,
-                                    25.101511419940255
-                                ],
-                                [
-                                    51.54583215335824,
-                                    25.101501436533947
-                                ],
-                                [
-                                    51.5458431173585,
-                                    25.101491372103187
-                                ],
-                                [
-                                    51.54583206437632,
-                                    25.10148138869682
-                                ],
-                                [
-                                    51.545832019885395,
-                                    25.10147136477823
-                                ],
-                                [
-                                    51.54582096690593,
-                                    25.101461381370967
-                                ],
-                                [
-                                    51.545809958417344,
-                                    25.10146142188148
-                                ],
-                                [
-                                    51.54579890544054,
-                                    25.101451438472516
-                                ],
-                                [
-                                    51.54579694798406,
-                                    25.101010386037615
-                                ],
-                                [
-                                    51.54580791194429,
-                                    25.101000321609483
-                                ],
-                                [
-                                    51.54579685900986,
-                                    25.10099033819904
-                                ],
-                                [
-                                    51.54579650311403,
-                                    25.100910146844193
-                                ],
-                                [
-                                    51.54578545018876,
-                                    25.100900163432627
-                                ],
-                                [
-                                    51.545774441750346,
-                                    25.100900203939663
-                                ],
-                                [
-                                    51.54575233590694,
-                                    25.100880237112264
-                                ],
-                                [
-                                    51.5457632998595,
-                                    25.100870172687383
-                                ],
-                                [
-                                    51.54576325537541,
-                                    25.10086014876786
-                                ],
-                                [
-                                    51.54575220245738,
-                                    25.10085016535362
-                                ],
-                                [
-                                    51.54576316640732,
-                                    25.10084010092876
-                                ],
-                                [
-                                    51.54575211349112,
-                                    25.10083011751447
-                                ],
-                                [
-                                    51.54576307743931,
-                                    25.10082005308961
-                                ],
-                                [
-                                    51.54575202452496,
-                                    25.10081006967525
-                                ],
-                                [
-                                    51.54576298847141,
-                                    25.10080000525041
-                                ],
-                                [
-                                    51.5457629439875,
-                                    25.10078998133078
-                                ],
-                                [
-                                    51.545751891075895,
-                                    25.10077999791633
-                                ],
-                                [
-                                    51.54577381896176,
-                                    25.10075986906587
-                                ],
-                                [
-                                    51.54578482738763,
-                                    25.1007598285591
-                                ],
-                                [
-                                    51.545795791326945,
-                                    25.10074976413188
-                                ],
-                                [
-                                    51.545795746840454,
-                                    25.100739740212244
-                                ],
-                                [
-                                    51.54580671077711,
-                                    25.10072967578424
-                                ],
-                                [
-                                    51.5458177192002,
-                                    25.100729635275062
-                                ],
-                                [
-                                    51.54582868313415,
-                                    25.10071957084546
-                                ],
-                                [
-                                    51.545828638645034,
-                                    25.100709546925838
-                                ],
-                                [
-                                    51.545850566505784,
-                                    25.10068941806427
-                                ],
-                                [
-                                    51.545883591763946,
-                                    25.100689296524717
-                                ],
-                                [
-                                    51.54589464467773,
-                                    25.100699279929437
-                                ],
-                                [
-                                    51.5458946891722,
-                                    25.10070930384896
-                                ],
-                                [
-                                    51.545916795006995,
-                                    25.100729270655734
-                                ],
-                                [
-                                    51.54592780342985,
-                                    25.10072923013839
-                                ],
-                                [
-                                    51.545938856350816,
-                                    25.100739213539637
-                                ],
-                                [
-                                    51.54592789242434,
-                                    25.100749277977233
-                                ],
-                                [
-                                    51.545938945347146,
-                                    25.100759261378414
-                                ],
-                                [
-                                    51.54596096219812,
-                                    25.100759180339534
-                                ],
-                                [
-                                    51.54597201512447,
-                                    25.100769163738185
-                                ],
-                                [
-                                    51.54597210412632,
-                                    25.100789211576767
-                                ],
-                                [
-                                    51.54596114019825,
-                                    25.10079927601678
-                                ],
-                                [
-                                    51.54597219312827,
-                                    25.100809259415307
-                                ],
-                                [
-                                    51.545972371132464,
-                                    25.1008493550922
-                                ],
-                                [
-                                    51.54599447700528,
-                                    25.100869321886385
-                                ],
-                                [
-                                    51.546005485440524,
-                                    25.100869281363064
-                                ],
-                                [
-                                    51.5460165383805,
-                                    25.100879264758024
-                                ],
-                                [
-                                    51.546005574448245,
-                                    25.100889329201273
-                                ],
-                                [
-                                    51.54601662739005,
-                                    25.10089931259616
-                                ],
-                                [
-                                    51.54600566345605,
-                                    25.100909377039418
-                                ],
-                                [
-                                    51.54601671639971,
-                                    25.100919360434247
-                                ],
-                                [
-                                    51.54600575246396,
-                                    25.100929424877517
-                                ],
-                                [
-                                    51.546005796967954,
-                                    25.10093944879654
-                                ],
-                                [
-                                    51.54601684991437,
-                                    25.10094943219128
-                                ],
-                                [
-                                    51.546005885976,
-                                    25.100959496634562
-                                ],
-                                [
-                                    51.54601693892426,
-                                    25.100969480029224
-                                ],
-                                [
-                                    51.54600597498415,
-                                    25.100979544472523
-                                ],
-                                [
-                                    51.546017027934255,
-                                    25.10098952786712
-                                ],
-                                [
-                                    51.546006063992394,
-                                    25.10099959231042
-                                ],
-                                [
-                                    51.54600610849655,
-                                    25.10100961622936
-                                ],
-                                [
-                                    51.546017161449406,
-                                    25.10101959962386
-                                ],
-                                [
-                                    51.54600619750494,
-                                    25.10102966406718
-                                ],
-                                [
-                                    51.54602830341614,
-                                    25.101049630855197
-                                ],
-                                [
-                                    51.54605032031878,
-                                    25.101049549802177
-                                ],
-                                [
-                                    51.546050364826634,
-                                    25.101059573720963
-                                ],
-                                [
-                                    51.54603940088143,
-                                    25.1010696381667
-                                ],
-                                [
-                                    51.54605045384244,
-                                    25.10107962155851
-                                ],
-                                [
-                                    51.54605058736632,
-                                    25.10110969331471
-                                ],
-                                [
-                                    51.546061640331885,
-                                    25.101119676705554
-                                ],
-                                [
-                                    51.54608365724694,
-                                    25.101119595647376
-                                ],
-                                [
-                                    51.546094710216046,
-                                    25.10112957903569
-                                ],
-                                [
-                                    51.54608374626841,
-                                    25.10113964348465
-                                ],
-                                [
-                                    51.54608379077918,
-                                    25.10114966740326
-                                ],
-                                [
-                                    51.54609484375105,
-                                    25.10115965079147
-                                ],
-                                [
-                                    51.5460838798008,
-                                    25.101169715240445
-                                ],
-                                [
-                                    51.54609493277451,
-                                    25.101179698628588
-                                ],
-                                [
-                                    51.5460839688225,
-                                    25.10118976307757
-                                ],
-                                [
-                                    51.54609502179806,
-                                    25.101199746465657
-                                ],
-                                [
-                                    51.54608405784431,
-                                    25.10120981091465
-                                ],
-                                [
-                                    51.546084102355245,
-                                    25.101219834833167
-                                ],
-                                [
-                                    51.54609515533357,
-                                    25.101229818221153
-                                ],
-                                [
-                                    51.54608419137719,
-                                    25.101239882670164
-                                ],
-                                [
-                                    51.54609524435736,
-                                    25.101249866058083
-                                ],
-                                [
-                                    51.54608428039924,
-                                    25.101259930507094
-                                ],
-                                [
-                                    51.54609533338124,
-                                    25.101269913894964
-                                ],
-                                [
-                                    51.54608436942138,
-                                    25.10127997834398
-                                ],
-                                [
-                                    51.54608459197714,
-                                    25.10133009793596
-                                ],
-                                [
-                                    51.5460956449656,
-                                    25.101340081323592
-                                ],
-                                [
-                                    51.54608468099962,
-                                    25.101350145772653
-                                ],
-                                [
-                                    51.5460626640433,
-                                    25.101350226831677
-                                ],
-                                [
-                                    51.54605170007369,
-                                    25.10136029127835
-                                ],
-                                [
-                                    51.546052100651984,
-                                    25.10145050654317
-                                ],
-                                [
-                                    51.54604113667272,
-                                    25.101460570989087
-                                ],
-                                [
-                                    51.546030128184604,
-                                    25.101460611515943
-                                ],
-                                [
-                                    51.54601916420262,
-                                    25.10147067596027
-                                ],
-                                [
-                                    51.5460192087088,
-                                    25.101480699878547
-                                ],
-                                [
-                                    51.54600824472414,
-                                    25.10149076432208
-                                ],
-                                [
-                                    51.54601929772122,
-                                    25.101500747715043
-                                ],
-                                [
-                                    51.546008333734825,
-                                    25.101510812158594
-                                ],
-                                [
-                                    51.5460084227456,
-                                    25.101530859995055
-                                ],
-                                [
-                                    51.54595338027303,
-                                    25.101531062609553
-                                ],
-                                [
-                                    51.545942416278486,
-                                    25.101541127048314
-                                ],
-                                [
-                                    51.54594246077855,
-                                    25.10155115096662
-                                ],
-                                [
-                                    51.545931496781336,
-                                    25.101561215404605
-                                ],
-                                [
-                                    51.545909479786665,
-                                    25.10156129644153
-                                ],
-                                [
-                                    51.54589851578582,
-                                    25.101571360877113
-                                ],
-                                [
-                                    51.54589864927553,
-                                    25.101601432632094
-                                ],
-                                [
-                                    51.545887685270266,
-                                    25.1016114970669
-                                ],
-                                [
-                                    51.54589873826878,
-                                    25.101621480468676
-                                ],
-                                [
-                                    51.54589878276545,
-                                    25.101631504386948
-                                ],
-                                [
-                                    51.545876765758074,
-                                    25.10163158541922
-                                ]
-                            ]
-                        ]
-                    },
-                    "properties": {
-                        "label": "building",
-                        "score": 0.375865638256073,
-                        "area": 2217.7800000065904,
-                        "shape_score": 1.0
-                    }
-                }""";
-
-        String feature2 = """
-                {
-                    "type": "Feature",
-                    "geometry": {
-                        "type": "Polygon",
-                        "coordinates": [
-                            [
-                                [
-                                    51.55302947581437,
-                                    25.105875342820088
-                                ],
-                                [
-                                    51.553018421858305,
-                                    25.105865359971794
-                                ],
-                                [
-                                    51.55301837677267,
-                                    25.10585533607124
-                                ],
-                                [
-                                    51.55300732281932,
-                                    25.105845353222044
-                                ],
-                                [
-                                    51.55299631395155,
-                                    25.105845394272595
-                                ],
-                                [
-                                    51.55298526000087,
-                                    25.1058354114217
-                                ],
-                                [
-                                    51.552985214918,
-                                    25.105825387521055
-                                ],
-                                [
-                                    51.552963107023864,
-                                    25.105805421816584
-                                ],
-                                [
-                                    51.552952098159594,
-                                    25.105805462863795
-                                ],
-                                [
-                                    51.55294104421609,
-                                    25.105795480009423
-                                ],
-                                [
-                                    51.55294099913691,
-                                    25.10578545610865
-                                ],
-                                [
-                                    51.55295200799941,
-                                    25.105785415062293
-                                ],
-                                [
-                                    51.55296297178092,
-                                    25.10577535011437
-                                ],
-                                [
-                                    51.55297398064247,
-                                    25.105775309066395
-                                ],
-                                [
-                                    51.552984944421276,
-                                    25.105765244116885
-                                ],
-                                [
-                                    51.55298485425589,
-                                    25.105745196315382
-                                ],
-                                [
-                                    51.55300687197348,
-                                    25.105745114215466
-                                ],
-                                [
-                                    51.5530178357469,
-                                    25.105735049263572
-                                ],
-                                [
-                                    51.55301774557628,
-                                    25.10571500146211
-                                ],
-                                [
-                                    51.553028754432326,
-                                    25.105714960410165
-                                ],
-                                [
-                                    51.55303971820128,
-                                    25.10570489545669
-                                ],
-                                [
-                                    51.553105771331516,
-                                    25.105704649123073
-                                ],
-                                [
-                                    51.55311682527982,
-                                    25.1057146319652
-                                ],
-                                [
-                                    51.5531168703732,
-                                    25.10572465586577
-                                ],
-                                [
-                                    51.55313897827704,
-                                    25.105744621547352
-                                ],
-                                [
-                                    51.55322704914414,
-                                    25.10574429303657
-                                ],
-                                [
-                                    51.553238103105734,
-                                    25.105754275869376
-                                ],
-                                [
-                                    51.55322713934901,
-                                    25.105764340837254
-                                ],
-                                [
-                                    51.55323819331245,
-                                    25.10577432366999
-                                ],
-                                [
-                                    51.55323828351926,
-                                    25.10579437147055
-                                ],
-                                [
-                                    51.55322731975904,
-                                    25.105804436438447
-                                ],
-                                [
-                                    51.553227364861606,
-                                    25.105814460338706
-                                ],
-                                [
-                                    51.55323841882966,
-                                    25.10582444317129
-                                ],
-                                [
-                                    51.55323846393316,
-                                    25.105834467071507
-                                ],
-                                [
-                                    51.55322745506681,
-                                    25.105834508139203
-                                ],
-                                [
-                                    51.55321649130217,
-                                    25.105844573106314
-                                ],
-                                [
-                                    51.553205482434876,
-                                    25.105844614172394
-                                ],
-                                [
-                                    51.55319451866752,
-                                    25.105854679137916
-                                ],
-                                [
-                                    51.553194608867535,
-                                    25.10587472693841
-                                ],
-                                [
-                                    51.55302947581437,
-                                    25.105875342820088
-                                ]
-                            ]
-                        ]
-                    },
-                    "properties": {
-                        "label": "building",
-                        "score": 0.4890393316745758,
-                        "area": 396.7362000034528,
-                        "shape_score": 1.0
-                    }
-                }""";
-
-        String feature3 = """
-                {
-                    "type": "Feature",
-                    "geometry": {
-                        "type": "Polygon",
-                        "coordinates": [
-                            [
-                                [
-                                    51.551290073990785,
-                                    25.105881818939412
-                                ],
-                                [
-                                    51.55127902017269,
-                                    25.105871835959075
-                                ],
-                                [
-                                    51.55127897522885,
-                                    25.10586181205559
-                                ],
-                                [
-                                    51.55126792141347,
-                                    25.105851829074354
-                                ],
-                                [
-                                    51.55121287705535,
-                                    25.105852033673408
-                                ],
-                                [
-                                    51.55120182324614,
-                                    25.105842050687126
-                                ],
-                                [
-                                    51.55120177830864,
-                                    25.10583202678347
-                                ],
-                                [
-                                    51.55117967069744,
-                                    25.105812060808226
-                                ],
-                                [
-                                    51.55113563522458,
-                                    25.105812224462657
-                                ],
-                                [
-                                    51.551124581425164,
-                                    25.105802241470396
-                                ],
-                                [
-                                    51.55112440170085,
-                                    25.105762145854943
-                                ],
-                                [
-                                    51.55113536563272,
-                                    25.105752081039594
-                                ],
-                                [
-                                    51.55117940108404,
-                                    25.105751917385604
-                                ],
-                                [
-                                    51.55120132893503,
-                                    25.105731787746134
-                                ],
-                                [
-                                    51.551201283997806,
-                                    25.105721763842325
-                                ],
-                                [
-                                    51.55121224791973,
-                                    25.10571169902141
-                                ],
-                                [
-                                    51.55126729221504,
-                                    25.105711494423648
-                                ],
-                                [
-                                    51.55127825613056,
-                                    25.10570142959794
-                                ],
-                                [
-                                    51.551278211187125,
-                                    25.105691405694216
-                                ],
-                                [
-                                    51.55128917509998,
-                                    25.10568134086772
-                                ],
-                                [
-                                    51.55133321052483,
-                                    25.105681177168496
-                                ],
-                                [
-                                    51.55134426432978,
-                                    25.10569116014526
-                                ],
-                                [
-                                    51.55133330042065,
-                                    25.105701224975757
-                                ],
-                                [
-                                    51.551344354227446,
-                                    25.105711207952467
-                                ],
-                                [
-                                    51.551399398521305,
-                                    25.105711003305682
-                                ],
-                                [
-                                    51.551421506149005,
-                                    25.105730969248135
-                                ],
-                                [
-                                    51.5514215511042,
-                                    25.105740993151556
-                                ],
-                                [
-                                    51.551432604921665,
-                                    25.10575097612145
-                                ],
-                                [
-                                    51.55152067581872,
-                                    25.1057506486239
-                                ],
-                                [
-                                    51.551531729644964,
-                                    25.10576063158624
-                                ],
-                                [
-                                    51.5515319095021,
-                                    25.105800727198947
-                                ],
-                                [
-                                    51.55152094559904,
-                                    25.105810792043066
-                                ],
-                                [
-                                    51.55143287465891,
-                                    25.10581111954151
-                                ],
-                                [
-                                    51.551421910746676,
-                                    25.105821184378453
-                                ],
-                                [
-                                    51.55142195570209,
-                                    25.10583120828176
-                                ],
-                                [
-                                    51.5514000278705,
-                                    25.105851337953272
-                                ],
-                                [
-                                    51.55134498351382,
-                                    25.105851542601354
-                                ],
-                                [
-                                    51.551334019590726,
-                                    25.10586160743193
-                                ],
-                                [
-                                    51.551345073412264,
-                                    25.10587159040812
-                                ],
-                                [
-                                    51.55133410948742,
-                                    25.1058816552387
-                                ],
-                                [
-                                    51.551290073990785,
-                                    25.105881818939412
-                                ]
-                            ]
-                        ]
-                    },
-                    "properties": {
-                        "label": "building",
-                        "score": 0.4893834888935089,
-                        "area": 508.857299995914,
-                        "shape_score": 1.0
-                    }
-                }""";
-
-        String doc = """
-                {
-                    "name": "mmsegmentation-loveda-b segmentations",
-                    "type": "FeatureCollection",
-                    "item_id": "20201211_223832_CS2",
-                    "crs": {
-                        "type": "name",
-                        "properties": {
-                            "name": "urn:ogc:def:crs:OGC:1.3:CRS84"
-                        }
-                    },
-                    "properties": {
-                        "tile_id": "20201211_223832_CS2",
-                        "model": {
-                            "id": "mmsegmentation-loveda-b",
-                            "params": {
-                                "rotation": 0
-                            }
-                        }
-                    },
-                    "features": [""" + feature1 + "," + feature2 + "," + feature3 + "]}";
-
-        return doc;
-    }
 }
