+++ src/main/java/com/datastax/oss/cass_stac/controller/IndexController.java
@@ -1,4 +1,4 @@
-package us.anant.cass.stac.controller;
+package com.datastax.oss.cass_stac.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,9 +6,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.datastax.oss.cass_stac.dto.CatalogDto;
+import com.datastax.oss.cass_stac.service.IndexService;
+
 import lombok.RequiredArgsConstructor;
-import us.anant.cass.stac.dto.CatalogDto;
-import us.anant.cass.stac.service.IndexService;
 
 @RestController
 @RequestMapping("/")
