+++ src/main/java/com/datastax/oss/cass_stac/dto/ItemDto.java
@@ -1,18 +1,21 @@
-package us.anant.cass.stac.dto;
+package com.datastax.oss.cass_stac.dto;
 
 import java.util.List;
+import java.util.Map;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Builder
 @Data
-public class CatalogDto {
+public class ItemDto {
 	private String type;
 	private String stac_version;
 	private List<String> stac_extensions;
 	private String id;
-	private String title;
-	private String description;
+	private GeoDto geometry;
+	private List<Float> bbox;
+	private String collection;
 	private List<LinkDto> links;
+	private Map<String, AssetDto> assets;
 }
