--- src/main/java/us/anant/cass/stac/StacapiApplication.java
@@ -1,13 +1,0 @@
-package us.anant.cass.stac;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StacapiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StacapiApplication.class, args);
-	}
-
-}
