--- src/main/resources/config.properties
@@ -1,40 +1,0 @@
-
-## Astra/Cassandra Cluster
-cassandra.keyspace=stac
-
-# Astra Connection
-#cassandra.secureBundlePath=
-#cassandra.password=
-
-# Cassandra Connection
-#cassandra.contactPoints=
-#cassandra.port=9042
-#cassandra.username=cassandra
-#cassandra.password=cassandra
-
-## Database Configuration
-database.partition.resolution.geo=6
-database.partition.resolution.time=MONTH
-database.partition.id_bucket_count=
-database.parallelism.read=10
-database.parallelism.write=10
-
-## Dao Configuration
-geojson.coordinateDecimalPlaces=16
-dao.item.property.DateFields=datetime,start_datetime,end_datetime,created,updated
-dao.item.property.IndexList.text=
-dao.item.property.IndexList.number=eo:cloud_cover,
-dao.item.property.IndexList.boolean=
-dao.item.property.IndexList.timestamp=datetime,
-
-dao.featureCollection.property.itemIdName=item_id
-
-dao.feature.batchSize=50
-dao.feature.label.property_name=label
-dao.feature.label.missingValue=unclassified
-dao.feature.property.IndexList.text=
-dao.feature.property.IndexList.number=
-dao.feature.property.IndexList.boolean=
-dao.feature.property.IndexList.timestamp=
-
-
